Cast diaSourceId as np.int64

+++ apps/utils.py
@@ -19,6 +19,7 @@ import gzip
 import yaml
 
 import pandas as pd
+import numpy as np
 from pyarrow import fs
 import pyarrow.parquet as pq
 import pyarrow.dataset as ds
@@ -155,7 +156,7 @@ def format_and_send_cutout_from_lsst(payload: dict) -> pd.DataFrame:
             }
             return Response(str(rep), 400)
 
-    filters = ds.field("diaSource", "diaSourceId") == int(payload["diaSourceId"])
+    filters = ds.field("diaSource", "diaSourceId") == np.int64(payload["diaSourceId"])
 
     args = yaml.load(open("config.yml"), yaml.Loader)
     hdfs = fs.HadoopFileSystem(args["HDFS"], args["HDFSPORT"], user=args["HDFSUSER"])
